refactor(api): extract report generation into shared service
